bloc-6/Traffic signs detection/apps/perf_bench/requirements.txt
streamlit
opencv-python-headless
Pillow
ultralytics
ncnn
tensorflow
